Add submission button to trigger auto PR
This adds a button to the checklist page, which sends the content of the
filled out form as an issue to the proceedings page. Submitters will be
required to have a Github account in order to submit the checklist,
allowing us to be able to be in contact with the user and have an idea
of who submitted the issue.

If a software name is not provided, an alert window pops up to inform
the user and prevent submission. If the checklist is valid (e.g. has a
name), then upon successful submission, the new issue page is opened
with the checklist template prefilled with the name and contents, while
the checklist page refreshes.